r-efi: mark basic types as Copy+Clone
Mark all the exported C-style enums as Copy+Clone, as well as the types
where it makes sense.

We don't, so far, set this on the other basic structures. Especially in
case of VLAs, rust does not even allow this. We can always extend this
in the future, if we agree on what to mark as Copy.

In case of the C-style enums it should be straightforward, though.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>