guid: clarify GUID endianness
Unlike RFC4122, the UEFI spec (Appendix A) requires the first three
fields of a GUID struct to be stored in LE byte-order. This removes
the need for flipping bytes and we can replace the constructors
`from_native()` and `from_spec()` with one, simply called
`from_fields()`, as they would now be the same.

Signed-off-by: Tom Gundersen <teg@jklm.no>