system/runtime: mark read-only arguments as *const
So far all pointers in the API were marked as *mut, this was just
as a safe default until we reviewed the spec in more detail.

Change Runtime Services to mark read-only argumnets as *const. Note
that this does not change the behavior, but it makes the code a bit
more self-documenting.

Signed-off-by: Tom Gundersen <teg@jklm.no>