base: convert booleans to newtype(u8)
Even though UEFI defines booleans to be 1-byte values that can be 0 or
1, all existing implementations use the C logic where anything but 0 is
true.

This commit changes our `Boolean` ABI to be a real `u8` type, so we can
actually deal with the entire range. Previously, if you transmuted a
UEFI boolean that is anything but 0 or 1 into our boolean, it would not
map to a valid enum-variant, and thus fall through all our match trees.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>