r-efi: sync 'const' annotations with spec
The spec clearly marks constant input with 'CONST'. It is not very
good in actually using it, but it does have the concept and applies it
to some of the argments.

The aim of this project is to provide a 1-to-1 mapping to the spec, so
lets make sure we do exactly that.

High-level wrappers should really re-evaluate whether the input they
pass is constant or not. However, it is not our job to judge whether
the function input actually modifies the data or not. This must be done
in layer that provides a safe wrapper, and it better be done with a
good justification.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>