base: alias Char8/Char16 with u8/u16 again
Lets move the unicode handling out of r-efi. We want r-efi to provide
constants, structures, symbols, types, and data from the spec, but we
do not want to mandate how it is to be employed. r-efi serves as base
layer to allow more high-level abstractions to no longer have to copy
information out of the spec. This does not imply that others do not
have to read the spec, though! You still have to know how each of these
functions behave.

This also drops the last unstable feature. From now on, r-efi can be
compiled on the stable channel.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>