build: bump version
Prepare the project for the next release and bump the version number.

We changed some function parameters that were incorrectly copied from
the spec. This is technically a backwards-incompatible change, which is
why we update the major number. It is unlikely that anyone used it so
far, but lets be sure.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>