examples: add uefi-cross-compile example
This imports another example called `uefi-cross-compile`. It is a
verbatim copy from the upstream rust-lang codegen test. We keep it in
our repository so we can run nightly tests on it and get notified of
possible breakage.

If you want to adjust this example, please make sure to file PRs against
upstream rust-lang as well.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>