protocol/hii_font: split hii_font and hii_font_ex
Keep both protocols separate just like the specification does it. This
keeps them better aligned with the specification. Downside is that the
naming scheme will diverge from the specification, but that is what the
`efi::` namespace is for, in case we ever want to update it to include
all names verbatim.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>