ci: document and use -Zbuild-std
Document the official alternative to `cargo-xbuild`, which is the
`-Zbuild-std` flag of cargo. Make the CI use it alongside `xbuild` so we
keep testing both.

Note that this is still marked as unstable and the official docs say:

    There are a large number of known issues with this implementation,
    and it is not intended to work for all targets and should not be
    used for anything other than experimentation and testing.

However, the `cargo-xbuild` and `xargo` toolchains had the same issues
in the past, so there is little reason not to give `-Zbuild-std` a try.

Thanks to Alex for figuring all this out!

Suggested-by: Alex James <theracermaster@gmail.com>
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>