system: verify MemoryType ABI
The MemoryType definition allows for a huge range of implementation
defined variants (all with the MSB set). We cannot declare a symbol for
each value, but we still need to support accepting such values. Until we
find a suitable way to represent this in rust, make sure

    `(MemoryType)v as u32`

is well-defined and works. This will allow us to seemlessly convert
between MemoryType and u32 representations and avoid restricting the
caller.

Reported-by: Richard Wiedenhöft <richard@wiedenhoeft.xyz>
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>