r-efi: convert enums to constants
This converts all `enum` types to `i32/u32` aliases with their variants
as constants. This is an API break, but the ABI stays unchanged.

The UEFI specification defines some enums that combine classic
enumerations with flags. One example is `MemoryType`, which reserves a
high-range with the MSB set for platform-specific types. With our
current implementation, we cannot reasonably represent those. Rust does
not allow undefined discrimants even for `repr(C)` enums. At the same
time, we cannot reasonable define an enum variant for all 2 billion
reserved identifiers.

Note that the UEFI specification is very clear on the ABI:

    Element of a standard ANSI C enum type declaration. Type INT32 or
    UINT32.  ANSI C does not define the size of sign of an enum so they
    should never be used in structures. ANSI C integer promotion rules
    make INT32 or UINT32 interchangeable when passed as an argument to
    a function.

Therefore, all enums should use a 32-bit underlying type and never be
used for structures. Promotion rules also guarantee that they are passed
as u32 to functions, so no worries there. For all intents and purposes,
we can thus treat them as `u32`.

There is one problem where the specification would extend an existing
structure with a negative variant, thus causing the underlying type to
switch to `i32`. With an `repr(C)`-enum (or with a C enum), this would
not be an API break, but it would be an API break if we use `u32`.
However, this sounds highly unlikely to happen, and even then would only
be a minor inconvenience. Even the type-change in a `repr(C)`-enum (and
more importantly in a C enum) would be a major nuissance and likely
break assumptions of existing applications. So we do not expect this to
happen.

Now there is several ways to transpose the C enums into Rust, each with
up- and downsides:

 * enum: By using rust enums, we get more type-safety, while retaining
   the possibility to extend the type with our own helpers. This allows
   to `match` on an enum and get completeness-checks by the compiler.
   Furthermore, enums and integers cannot be mixed up unintentionally.
   The downside are the aforementioned problems of flags mixed in an
   enum. We would be forced to use `u32` in all function-definitions
   where the enum-type is expected, otherwise users would be unable to
   pass the right values. An alternative is to provide additional `*Raw`
   function calls alongside the typed-ones, which use `u32`.
   Using `u32` in function definitions is likely to trip us in the
   future when we forget this when transposing new data from the spec.
   At the same time, it requires every caller to convert the enum to an
   u32 at the call-site, which just feels like a mismatch.

 * newtype: By using the newtype-idiom we could wrap a u32 into a
   structure and put `repr(transparent)` on it. This would gives us free
   hands at properly wrapping every type, but it would require lots of
   policy decisions and feels out-of-scope of this project. It would no
   longer be a straightforward transpose of the specification.

 * alias: By aliasing these types as `u32` we get a straightforward
   transpose from the spec, and direct mapping to the C behavior.
   However, it is the least ergonomic solution and provides no real
   benefits over C.

This patch chooses the `alias` option and removes all Rust `enum`s from
the code-base. The reason for this is that the UEFI specification itself
chose this option for a lot of the newer protocols. It avoids all the
ABI nastiness around enums and easy extensions in the future.

With the nature of this project in mind, it seems to be the right
solution to provide the raw constants with a clear and simple mapping
from the spec, rather than the more ergonomic solutions. Higher-level
crates are likely to wrap every single lower type, anyway. And with this
solution, they retain the full freedom to do this, while we take care of
transposing the spec and guaranteeing correctness of the underlying
values, and monitoring the spec for changes.

Thanks to Richard for reporting the issue and evaluating different
options how to solve this!

Reported-by: Richard Wiedenhöft <richard@wiedenhoeft.xyz>
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>