r-efi: convert incomplete types to empty arrays
Convert all incomplete structure types (which use an unbound trailing
array) to a complete type by explicitly sizing the array as 0. This has
the side-effect of avoiding making the type a DST.

The reason to do this is that DSTs have fat-pointers in rust, even for
raw-pointers. Those violate the UEFI ABI and thus we cannot use them.
Using empty arrays seems like the most simple solution for now. This
might use more elegant solutions in the future, but this works for now.

Reported-by: GGRei
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>