build: setup for std dependency
Rustc requires special dependency setups for crates from crates.io to
be built as dependency of rustc (e.g., core, std, ...). Special dummy
crates on crates.io provide the hooks that the build-system can then
override. All we have to do is to make sure to pull in those
dependencies when built as part of rustc.

Signed-off-by: Ayush Singh <ayushsingh1325@gmail.com>
(Fix alphabetical order and drop reference to non-existant rustc path.)
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>