protocol/device_path_to_text: mark pointers as *mut
The returned pointers are mutable, according to the spec. Note that they
are allocated and it is the responsibility of the caller to release
them, so it actually makes sense to mark them mutable.

For the protocol pointer, we want consistency with the other protocols.
We did not apply `*const` so far, since the protocol implementations
violate it anyway. We so far left it to the safe-wrappers to deal with
this and convert appropriately.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>