r-efi: convert ZSTs to const-generics
Convert all trailing-struct ZSTs to const-generics with a default of 0.
The default guarantess that the types stay backwards-compatible, while
the new const-generics allow compile-time adjustments of those trailing
arrays.

This does not solve the problem of accessing trailing-struct ZSTs with
runtime-dynamic sizes, but it paves the way to better ZST handling in
r-efi.

Signed-off-by: Ayush Singh <ayushsingh1325@gmail.com>
(extended commit-message)
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>