base: reduce Guid alignment to 4-byte
Reduce the alignment requirement for `Guid` to 4-bytes. This goes
against the UEFI specification, but follows what EDK2 does:

https://github.com/tianocore/edk2/blob/master/BaseTools/Source/C/Include/Common/UefiBaseTypes.h#L18

This is a breaking change for pure r-efi based systems. But given that
it fixes interaction with EDK2 (and possibly other implementations), we
consider this a bug-fix.

Discussion to fix the specification has been initiated on the EDK2
board.

Reported-by: Nicholas Bishop <nbishop@nbishop.net>
Signed-off-by: David Rheinsberg <david@readahead.eu>