system: fix signature of BootUninstallMultipleProtocolInterfaces
The respective prototype takes a plain handle as first argument, rather
than a pointer to it. Fix this.

Note that the prototype currently cannot be used in Rust, given the lack
of support for variadic functions for `eficall`. However, this change
might still break code that purely deals with function pointers without
invoking them. However, since this fixes a bug in this implementation,
we will consider this covered by semver and avoid a major version bump.

Signed-off-by: Ayush Singh <ayushdevel1325@gmail.com>
(elaborate situation in commit-message)
Signed-off-by: David Rheinsberg <david@readahead.eu>